Squashed commit (from master) of the following:
commit cc2658e9df7c9c51158b408da57bf38b9d04a3d8
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Wed May 18 19:39:59 2022 +0300

    new version 0.5.02

commit 6e248f0e9e2ef3d364658521417fbfe5d357bd9d
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Wed May 18 19:35:31 2022 +0300

    new system of versions output

commit d07e651e21e1724647ea4003ef6d7f1243f90a66
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Wed May 18 19:08:26 2022 +0300

    Default DFT y_to increased

commit c282cc3c6847c3b6b6b94a81b8b076fdd3a78a79
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Wed May 18 15:36:40 2022 +0300

    exe ico changed

commit 956145a1fbb963d352e990df180d38d3c446c57b
Merge: 3e3b907 59172b3
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 23:38:54 2022 +0300

    Merge report->master

    - report update
    - CD update

commit 3e3b9072ec629273cdd5c9f46ae6d76b1f56baea
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 23:17:26 2022 +0300

    TODO.txt: all is done

commit 8419b07571147f5ca7931bdf322f65996275898f
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 23:16:42 2022 +0300

    btn name changed

commit 293ebe36d664a5528953ab498491a23d9870eb25
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 23:07:48 2022 +0300

    back conv inf fixed

commit 5e47eb86ec8dbc21340140d520b1b8d6d0356588
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:58:36 2022 +0300

    cb_anim set default checked

commit 631951f174e601f2665d2ad8f3a67bcf00de1f46
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:57:03 2022 +0300

    comment

commit 9551eb1279a2f7ec7ef8031b6e1861bd00644ce4
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:54:52 2022 +0300

    slider_step: resetting on button reset

commit f1477c9b4f2b754bf6698cf943634b3db12d47ac
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:51:28 2022 +0300

    __COURSE__VERSION macro renamed to __COURSE_VERSION

commit 99feb4e9e583c5ef6916a6bd77073568348def56
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:49:39 2022 +0300

    AboutDlg: fixed version No

commit 38db663bc1d0a9994633c23dcf74f5112a5d6e6d
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:33:40 2022 +0300

    ToDo update

commit e533be854ba6d10591f57470eda9def3097c9415
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:28:54 2022 +0300

    ToDo update

commit 8be553788325518ac1cae12cfd3e667e7b89fc7c
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Tue May 17 22:25:58 2022 +0300

    TODO created

commit f422af918d0e7750a4ac112b26a9cccaf6cf8a76
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 12:11:09 2022 +0300

    new version

commit b2f2ae639b9b1242767a1efb1d910a993d08f112
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 12:09:41 2022 +0300

    content table

commit d9985e057d8db47d9c08aac6ed446dc4ec867dd6
Merge: 1571298 caec328
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 09:31:12 2022 +0300

    Merge branch 'master' of https://github.com/Dimach24/coursework

commit 1571298902a6fc4162854cbafd7b9a9298f325f0
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 09:28:24 2022 +0300

    changes commented

commit 47fd9a57e5bb8387dfb7bc68f6d37eca5b636e43
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 09:25:24 2022 +0300

    fixed back conv

commit b551d14759e4fc789db2c9fecfc7fc89e38dd3d7
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 09:03:49 2022 +0300

    interface: calculator win name

commit e8cfcc9173ea896eac2bd05011f0bf1e5c96f14e
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 09:02:37 2022 +0300

    interface: tabbing

commit 43eddcac435173307eba46b30983a3e52f62f907
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Mon May 16 08:50:27 2022 +0300

    redesigned

commit caec328d04dff6ca4b07a430d9f288dacc1c6b7b
Author: D.mon 24 <66939631+Dimach24@users.noreply.github.com>
Date:   Sat May 14 10:31:03 2022 +0300

    Update README.md

commit a75e6046fc56e0ccfedb12836e0192dd6a3f0e80
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Sat May 14 10:27:51 2022 +0300

    file saving: comments

commit eb185d5f75a1de49fb1f92a5c1b87d57cbff2ceb
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Sat May 14 10:08:54 2022 +0300

    saving: offset and clearing up

commit 3969cb206cd13066c9f30222be688ee0def8e12d
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Sat May 14 09:47:23 2022 +0300

    correct saving

commit a4799ce62ac53d7e35e68fbace72790a731bd44a
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Thu May 12 18:01:57 2022 +0300

    sys menu save added

    and cassert included to the MF.cpp

commit 801101f5224f57810c33b0d7b0cc9475ddef2456
Author: D.mon 24 <dchaminov@ya.ru>
Date:   Wed May 11 15:27:16 2022 +0300

    idk

+++ coursework/CMyGraph.h
@@ -184,8 +184,8 @@ public:
 	/// <param name="wx">x coord in the elements coords</param>
 	/// <param name="wy">y coord in the elements coords</param>
 	/// <param name="rect">rectangular border of the drawing region (of this element)</param>
-	/// <returns>pair of float - real (math) coords</returns>
-	std::pair<float, float> dotToCoords(int wx, int wy, CRect rect);
+	/// <returns>pair of double - real (math) coords</returns>
+	std::pair<double, double> dotToCoords(int wx, int wy, CRect rect);
 
 	/// <summary>
 	/// Direct conversion real (math) to element (in pixels) coords
